requirements.txt
claude-agent-sdk
opencv-python
numpy
python-dotenv
pillow
pillow-heif
anyio>=4.4.0
